Add League Trawler
League trawler starts with a single user and recurses through
all users this user is in a league with. Given that most
sleeper users (on average) play in more than one league, it's
fairly likely this recursion should reach a large number of
leagues.

With this recursion, the aim is populate a trade database catalog
over the entirety of sleeper.